+++ src/main/java/com/wnis/linkyway/service/TrashService.java
@@ -1,11 +1,11 @@
 package com.wnis.linkyway.service;
 
-import com.wnis.linkyway.dto.card.CardResponse;
+import com.wnis.linkyway.dto.card.io.CardResponse;
 import com.wnis.linkyway.dto.tag.TagResponse;
 import com.wnis.linkyway.entity.Card;
 import com.wnis.linkyway.exception.common.NotFoundEntityException;
-import com.wnis.linkyway.repository.CardRepository;
-import com.wnis.linkyway.repository.CardTagRepository;
+import com.wnis.linkyway.repository.card.CardRepository;
+import com.wnis.linkyway.repository.cardtag.CardTagRepository;
 import com.wnis.linkyway.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
